Add files via upload
Created a Baseclass utilityfor for webdriver initialization, as this web driver initialization or closing is common and should be sharen among all the test cases.
TestLogin : Created a test method for Login multiple users. To verify the test method with multiple credentials/multiple sets of data with the help of DATA PROVIDER annotation.
Pom for recLogin : Page object model implementation for recruiter Login.